Bug fix #428: wrong parameter in OperationFailed message
In nfsmount function of nf_utils.py, OperationFailed was being
called as:

raise OperationFailed("GINFS00018E", err)

and was generating an "'str' object has no attribute 'iteritems'"
error because the function is expecting a dict instead of a
string. Changing it to:

raise OperationFailed("GINFS00018E", {'err': err})

Fixes the problem and now the correct error message is
displayed.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>