Removing the 'Server' UI tab from 'Host' super tab
Concerns were raised about the 'Server' UI, a remote feature,
being listed together with other local Ginger features underneath
the 'Host' section.

There's also a problem with how to hide the feature if the
'ipmitool' command isn't available. With other Ginger features,
it is trivial to hide the frame if the Capabilities API returns
'false' for a given feature. In this case, however, the whole
tab would need to be hid. This is not something we have predicted
in WoK and will require changes in both WoK and Ginger to happen,
which isn't something we're going to do at this moment.

Note that this patch simply erases the tab declaration from
tab-ext.xml - any pre-existent backend/UI code that belongs
to the feature is preserved.

Signed-off-by: Daniel Henrique Barboza <danielhb@linux.vnet.ibm.com>