Fix for confirmation message for delete partition
the message provided when deleting partition is dependent on vg,
i.e., if partition is part of vg then message displayed is -

'Selected Partition is part of a Volume group. Removal may hamper
the Volume group usage. Are you sure you want to delete this partition?'

if not, the confirmation message displayed is -
'Do you want to delete the partition?'

Signed-off-by: Suresh Babu Angadi <sureshab@in.ibm.com>