Fix expand/collpase widget configuration in Network tab
There are some elements which 2 settings for aria-expanded.
aria-expanded should be set to true when the widget is open (expanded) and,
in addition to that, the content should have the class "collapse in".
Otherwise, the expand icon will not reflect the right behavior.
When setting aria-expanded=false, only set the content class as "collapse".

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>